Split out router to separate file. Move app code to lib. Fix routing with maybe_patch. Fix nav styles.
