+++ frontend/index.html
@@ -14,9 +14,6 @@
                 font-weight: 500;
                 font-size: 13px;
             }
-            a, a:hover, a:visited {
-                color: #fff;
-            }
             *, *:before, *:after {
                 box-sizing: border-box;
             }
